https://github.com/RedRoverSchool/JenkinsQA_Java_2026_spring/issues/662
